requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
httpx>=0.25.2
requests>=2.31.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0
PyJWT>=2.10.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.16.0
databases>=0.9.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
loguru>=0.7.0
tenacity>=9.1.2
python-telegram-bot>=20.0
pytest>=7.4.0
pytest-cov>=4.1.0
protobuf>=4.25.3
# Основные зависимости для работы с изображениями и системами
Pillow>=11.2.1
psutil>=5.9.0
numpy<2.0.0

# Зависимости для машинного обучения и ИИ
torch>=2.7.1
xformers>=0.0.23.post1

# HTTP клиент для API запросов
aiohttp>=3.9.0

# Зависимости для аутентификации
email-validator>=2.3.0
dnspython>=2.8.0

# Зависимости для email
python-dotenv>=1.0.0

# Зависимости для OAuth
authlib>=1.2.0
httpx>=0.24.0

# Системные зависимости
GPUtil
psycopg2-binary
pytest-asyncio
aiosqlite>=0.19.0

# Зависимости для Yandex Cloud Storage
boto3>=1.34.0
botocore>=1.34.0

# Redis для кэширования
redis>=5.0.0
hiredis>=2.2.0

# Celery для фоновых задач
celery>=5.3.0
flower>=2.0.0